refactor: Imports of codemetropolis.toolchain.commons.blockmodifier.World have been corrected in the codemetropolis.toolchain.rendering.model.primitive.Boxel and codemetropolis.toolchain.rendering.control.WorldBuilder classes. Additionally, unnecessary output XMLs have been deleted.
BREAKING CHANGE: The program is in a non-executable state!
